standardized code of 3 layers
